celery-queue/requirements.txt
oss2==2.12.1
celery==3.1.25
thrift==0.9.3
Pillow==7.2.0
